Support writing 64-bit varints on 32-bit Python
Previously our varint encoding was limited to `SIZE_MAX`, which is
64-bit on 64-bit builds but 32-bit on 32-bit builds. Switch to
`uint64_t` instead, in order to explicitly allow 64-bit values to be
serialized even on 32-bit builds.

+++ src/memray/_memray/record_reader.cpp
@@ -102,7 +102,7 @@ RecordReader::readHeader(HeaderRecord& header)
 }
 
 bool
-RecordReader::readVarint(size_t* val)
+RecordReader::readVarint(uint64_t* val)
 {
     *val = 0;
     int shift = 0;
@@ -113,7 +113,7 @@ RecordReader::readVarint(size_t* val)
             return false;
         }
 
-        *val |= (static_cast<size_t>(next & 0x7f) << shift);
+        *val |= (static_cast<uint64_t>(next & 0x7f) << shift);
         if (0 == (next & 0x80)) {
             return true;
         }
@@ -126,14 +126,14 @@ RecordReader::readVarint(size_t* val)
 }
 
 bool
-RecordReader::readSignedVarint(ssize_t* val)
+RecordReader::readSignedVarint(int64_t* val)
 {
-    size_t zigzag_val;
+    uint64_t zigzag_val;
     if (!readVarint(&zigzag_val)) {
         return false;
     }
 
-    *val = static_cast<ssize_t>((zigzag_val >> 1) ^ (~(zigzag_val & 1) + 1));
+    *val = static_cast<int64_t>((zigzag_val >> 1) ^ (~(zigzag_val & 1) + 1));
     return true;
 }
